Add assertions
Assertions were added to prevent programmer errors from breaking the code. Files
that assertions were added to are Task.java, Todo.java, Event.java, Deadline.java,
TaskList.java, Storage.java, Praser.java and others.

An example of how it was added is in Event.java, where we asert to ensure that
the descriptiom, from and to fields are not empty. This will ensure the app
will not be broken by null values in these fields.